+++ make.py
@@ -4,7 +4,7 @@ import pypackager
 import os
 p=pypackager.PyPackager("habitjewel") #package name
 p.display_name = 'HabitJewel' #package display name in HAM
-p.version = '0.8.0' #package version
+p.version = '0.8.1' #package version
 p.buildversion = '0-local' #package build version
 p.description="""Record and track your progress in achieving your daily habits.""" #package description
 p.author='Ashley Hooper' #package author
@@ -31,9 +31,12 @@ chmod 755 /opt/habitjewel/habitjewel.py
 
 #p.postremove="""#!/bin/sh
 #"""
-p.changelog ="""## [0.8.0] - 2016-09-26
-### Changed
-- Fixed schema upgrade bug that caused habits table to be recreated without primary key
+p.changelog ="""## [0.8.1] - 2016-09-27
+### Added
+- try/except blocks to Day and Master habits lists to stop bailing out in case of
+  errors
+- code to fix habits with NULL ids (due to bug in earlier schema update code which
+  recreated habits table without id as PRIMARY KEY)
 """ # the changelog
 dir_name='src' #src directory name
 #Here we will loop in all files, directory in src to add it to pkg