+++ src/opt/habitjewel/CHANGELOG.md
@@ -2,6 +2,14 @@
 ### Added
 
 ### Changed
+
+## [0.8.1] - 2016-09-27
+### Added
+- try/except blocks to Day and Master habits lists to stop bailing out in case of
+  errors
+- code to fix habits with NULL ids (due to bug in earlier schema update code which
+  recreated habits table without id as PRIMARY KEY)
+
 ## [0.8.0] - 2016-09-26
 ### Changed
 - Fixed schema upgrade bug that caused habits table to be recreated without primary key